requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
discord.py==1.7.3
google==3.0.0
idna==3.2
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.1
urllib3==1.26.6
yarl==1.6.3
